Revert delegation of configuration constants to autoloader
