--- .autoloader.atoum.php
@@ -1,3 +1,0 @@
-<?php
-// Empty autoloader for atoum to prevent the usage of the GLPI autoloader
-// that defines some constants that must be redefined in the tests bootstrap file.
